+++ CSharp/DataParser.cs
@@ -12,19 +12,19 @@ namespace CSharp
       return System.IO.File.ReadAllLines(filename);
     }
 
-    public static Dictionary<int,Dictionary<int,float>> write2DArray(string filename, string split) {
+    public static Dictionary<int,Dictionary<int, double>> write2DArray(string filename, string split) {
       string[] data = writeArray(filename);
-      var result = new Dictionary<int,Dictionary<int,float>>();
-      Dictionary<int,float> score;
+      var result = new Dictionary<int,Dictionary<int, double>>();
+      Dictionary<int, double> score;
 
       foreach (string line in data) {
         string[] lineResult = line.Split(split);
-        score = new Dictionary<int, float>();
-        score.Add(int.Parse(lineResult[1]),  float.Parse(lineResult[2]));
+        score = new Dictionary<int, double>();
+        score.Add(int.Parse(lineResult[1]),  double.Parse(lineResult[2]));
 
         if ( result.TryAdd( int.Parse(lineResult[0]), score)) {
         } else {
-          result[int.Parse(lineResult[0])].Add(int.Parse(lineResult[1]), float.Parse(lineResult[2]) / 10);
+          result[int.Parse(lineResult[0])].Add(int.Parse(lineResult[1]), double.Parse(lineResult[2]) / 10);
         }
       }
 
@@ -33,24 +33,24 @@ namespace CSharp
 
 
 
-    public static Dictionary<int,Dictionary<int,float>> write2DArray_MovieLens(string filename) { 
+    public static Dictionary<int,Dictionary<int, double>> write2DArray_MovieLens(string filename) { 
       string[] data = writeArray(filename).Skip(1).ToArray(); //Skip first item because first line is definition as it is a csv.
-      var result = new Dictionary<int, Dictionary<int,float>>();
+      var result = new Dictionary<int, Dictionary<int, double>>();
       
       foreach (string line in data) {
         string[] lineResult = line.Split(',');
-        Dictionary<int,float> score = new Dictionary<int, float>();
-        score.Add(int.Parse(lineResult[1]), float.Parse(lineResult[2]));
+        Dictionary<int, double> score = new Dictionary<int, double>();
+        score.Add(int.Parse(lineResult[1]), double.Parse(lineResult[2]));
 
         if (result.TryAdd(int.Parse(lineResult[0]), score)) { 
         } else {
-          result[int.Parse(lineResult[0])].Add(int.Parse(lineResult[1]), float.Parse(lineResult[2]) / 10);
+          result[int.Parse(lineResult[0])].Add(int.Parse(lineResult[1]), double.Parse(lineResult[2]) / 10);
         }
       }
       return result;
     }
 
-    public static List<int> getProducts(Dictionary<int, Dictionary<int,float>> data) {
+    public static List<int> getProducts(Dictionary<int, Dictionary<int, double>> data) {
       List<int> result = new List<int>();
       foreach (var user in data) {
         foreach (var product in user.Value) {
@@ -63,8 +63,8 @@ namespace CSharp
       return result;
     }
 
-    public static Dictionary<int, Dictionary<int, float>> filterRatings(Dictionary<int, Dictionary<int, float>> data, int productId) {
-      Dictionary<int, Dictionary<int, float>> result = new Dictionary<int, Dictionary<int, float>>();
+    public static Dictionary<int, Dictionary<int, double>> filterRatings(Dictionary<int, Dictionary<int, double>> data, int productId) {
+      Dictionary<int, Dictionary<int, double>> result = new Dictionary<int, Dictionary<int, double>>();
       foreach (var user in data) {
         if (user.Value.ContainsKey(productId)) {
           result.Add(user.Key, user.Value);
@@ -73,8 +73,8 @@ namespace CSharp
       return result;
     }
     
-    public static Dictionary<int, Dictionary<int, float>> filterTwoProducts(Dictionary<int, Dictionary<int, float>> data, int productId1, int productId2) {
-      Dictionary<int, Dictionary<int, float>> result = new Dictionary<int, Dictionary<int, float>>();
+    public static Dictionary<int, Dictionary<int, double>> filterTwoProducts(Dictionary<int, Dictionary<int, double>> data, int productId1, int productId2) {
+      Dictionary<int, Dictionary<int, double>> result = new Dictionary<int, Dictionary<int, double>>();
       foreach (var user in data) {
         if (user.Value.ContainsKey(productId1) && user.Value.ContainsKey(productId2)) {
           result.Add(user.Key, user.Value);
@@ -83,8 +83,8 @@ namespace CSharp
       return result;
     }
 
-    public static Dictionary<int, Dictionary<int,float>> filterRatings(Dictionary<int, Dictionary<int, float>> data, int productId, int target) {
-      Dictionary<int, Dictionary<int,float>> result = new Dictionary<int, Dictionary<int, float>>();
+    public static Dictionary<int, Dictionary<int, double>> filterRatings(Dictionary<int, Dictionary<int, double>> data, int productId, int target) {
+      Dictionary<int, Dictionary<int, double>> result = new Dictionary<int, Dictionary<int, double>>();
       foreach (var user in data) {
         if (user.Key == target) {
           result.Add(user.Key, user.Value);
@@ -95,19 +95,19 @@ namespace CSharp
       return result;
     }
 
-    public static void checkData(Dictionary<int, Dictionary<int, float>> data) {
+    public static void checkData(Dictionary<int, Dictionary<int, double>> data) {
     }
 
-    public static Dictionary<int,Dictionary<int,float>> filterData(Dictionary<int,Dictionary<int,float>> data, int keyX, int keyY, bool delete=true) {
+    public static Dictionary<int,Dictionary<int, double>> filterData(Dictionary<int,Dictionary<int, double>> data, int keyX, int keyY, bool delete=true) {
       var firstUser = data[keyX];
       var secondUser = data[keyY];
       
-      var result = new Dictionary<int, Dictionary<int, float>>();
+      var result = new Dictionary<int, Dictionary<int, double>>();
       
       if (delete == true) {
 
-        var firstResult = new Dictionary<int, float>();
-        var secondResult = new Dictionary<int, float>();
+        var firstResult = new Dictionary<int, double>();
+        var secondResult = new Dictionary<int, double>();
 
         foreach (var firstValue in firstUser) {
           if ( secondUser.ContainsKey(firstValue.Key) == true) {
@@ -122,8 +122,8 @@ namespace CSharp
 
       } else {
 
-        var firstResult = new Dictionary<int,float>();
-        var secondResult = new Dictionary<int,float>();
+        var firstResult = new Dictionary<int, double>();
+        var secondResult = new Dictionary<int, double>();
         var keysChecked = new List<int>();
 
         foreach (var firstValue in firstUser) {
@@ -151,16 +151,16 @@ namespace CSharp
     }
 
     public static Tuple<
-      Dictionary<int,float>,
-      Dictionary<int,float>>
-      splitDictionaries (Dictionary<int,Dictionary<int,float>> data) {
+      Dictionary<int, double>,
+      Dictionary<int, double>>
+      splitDictionaries (Dictionary<int,Dictionary<int, double>> data) {
       //Acting under the pretense that every dataset this method gets
       //Has gone through the datafilter in DataParser
       //So we know it always has a matched pair of keys
       //And 2 comparees
         int n = 0;
-        var firstItem = new Dictionary<int, float>();
-        var secondItem = new Dictionary<int, float>();
+        var firstItem = new Dictionary<int, double>();
+        var secondItem = new Dictionary<int, double>();
         
         foreach (var dataItem in data) {
           //Console.WriteLine(dataItem.Key);
@@ -172,7 +172,7 @@ namespace CSharp
           n += 1;
         }
 
-        return new Tuple<Dictionary<int, float>, Dictionary<int, float>>(firstItem, secondItem);
+        return new Tuple<Dictionary<int, double>, Dictionary<int, double>>(firstItem, secondItem);
     }
   }
 }
