+++ CSharp/Neighbor.cs
@@ -7,9 +7,9 @@ namespace CSharp
     class Neighbor {
         public int userId;
         public double similarity;
-        public Dictionary<int,float> review;
+        public Dictionary<int, double> review;
 
-        public Neighbor(int userId, double similarity, Dictionary<int,float> review) {
+        public Neighbor(int userId, double similarity, Dictionary<int, double> review) {
             this.userId = userId;
             this.similarity = similarity;
             this.review = review;
